Remove XMLLineProcessor and move methods to XMLProcessor
